[CREATE QUOTA - BACKEND] Re-using of some stuff from **Create Measures V2**
[TRELLO CARD](https://trello.com/c/Kiospkws/361-dit-create-quota-iteration)
