+++ app/forms/workbasket_forms/create_measures_form.rb
@@ -1,0 +1,6 @@
+module WorkbasketForms
+  class CreateMeasuresForm < ::WorkbasketForms::BaseForm
+
+    attr_accessor :workbasket_name
+  end
+end
