+++ app/forms/workbasket_forms/create_quota_form.rb
@@ -1,0 +1,9 @@
+module WorkbasketForms
+  class CreateQuotaForm < ::WorkbasketForms::BaseForm
+
+    attr_accessor :order_number,
+                  :description,
+                  :licensed,
+                  :license_id
+  end
+end
