--- app/forms/workbaskets/create_quota_form.rb
@@ -1,9 +1,0 @@
-module Workbaskets
-  class CreateQuotaForm < ::Workbaskets::BaseForm
-
-    attr_accessor :order_number,
-                  :description,
-                  :licensed,
-                  :license_id
-  end
-end
