+++ app/interactors/workbasket_interactions/create_measures/submit_for_cross_check.rb
@@ -1,0 +1,6 @@
+module WorkbasketInteractions
+  module CreateMeasures
+    class SubmitForCrossCheck < ::WorkbasketInteractions::SubmitForCrossCheckBase
+    end
+  end
+end
