+++ app/interactors/workbasket_interactions/create_quota/submit_for_cross_check.rb
@@ -1,0 +1,6 @@
+module WorkbasketInteractions
+  module CreateQuota
+    class SubmitForCrossCheck < ::WorkbasketInteractions::SubmitForCrossCheckBase
+    end
+  end
+end
