--- app/interactors/workbaskets/create_measures/submit_for_cross_check.rb
@@ -1,6 +1,0 @@
-module Workbaskets
-  module CreateMeasures
-    class SubmitForCrossCheck < ::Workbaskets::SubmitForCrossCheckBase
-    end
-  end
-end
