--- app/interactors/workbaskets/create_quota/settings_saver.rb
@@ -1,21 +1,0 @@
-module Workbaskets
-  module CreateQuota
-    class SettingsSaver < ::Workbaskets::SettingsSaverBase
-
-      WORKBASKET_TYPE = "CreateQuota"
-
-      ASSOCIATION_LIST = %w(
-        quota_periods
-        conditions
-        footnotes
-        excluded_geographical_areas
-      )
-
-      ASSOCIATION_LIST.map do |name|
-        define_method("#{name}_errors") do |measure|
-          get_association_errors(name, measure)
-        end
-      end
-    end
-  end
-end
