--- app/interactors/workbaskets/create_quota/submit_for_cross_check.rb
@@ -1,6 +1,0 @@
-module Workbaskets
-  module CreateQuota
-    class SubmitForCrossCheck < ::Workbaskets::SubmitForCrossCheckBase
-    end
-  end
-end
