--- app/interactors/workbaskets/submit_for_cross_check_base.rb
@@ -1,25 +1,0 @@
-module Workbaskets
-  class SubmitForCrossCheckBase
-    attr_accessor :workbasket
-
-    def initialize(workbasket)
-      @workbasket = workbasket
-    end
-
-    def run!
-      workbasket.move_status_to!(:submitted_for_cross_check)
-      update_collection!
-    end
-
-    private
-
-      def update_collection!
-        workbasket.settings
-                  .collection
-                  .map do |item|
-
-          item.move_status_to!(:submitted_for_cross_check)
-        end
-      end
-  end
-end
