+++ app/services/workbasket_services/association_savers/base.rb
@@ -1,0 +1,14 @@
+module WorkbasketServices
+  module AssociationSavers
+    class Base
+
+      DEFAULT_LANGUAGE = "EN"
+
+      private
+
+        def set_primary_key(record)
+          ::WorkbasketValueObjects::Shared::PrimaryKeyGenerator.new(record)
+        end
+    end
+  end
+end
